refactor: post message bridge를 apps/provider로 이동
