+++ apps/webview/app/provider/RootProvider.tsx
@@ -2,7 +2,7 @@
 
 import { PropsWithChildren } from 'react'
 import ReactQueryProvider from './ReactQueryProvider'
-import { BridgeProvider } from '../../WebPostMessageBridge'
+import { BridgeProvider } from './PostMessageBridgeProvider'
 
 export const RootProvider = ({ children }: PropsWithChildren) => (
   <ReactQueryProvider>
