+++ apps/webview/app/provider/index.ts
@@ -1,3 +1,4 @@
 import { RootProvider } from './RootProvider'
 
+export * from './PostMessageBridgeProvider'
 export { RootProvider }
